Develop and finalize ETL pipeline script
I downloaded the data, categories.csv and messages.csv, to this project. The  "ETL Pipeline Preparation" Jupyter notebook to develop a script that cleans and joins the data in these files and writes the result to the table LabeledMessages in data.db. The script itself was then created as "etl_pipeline.py".