Mostly changes to ETL pipeline
I changed the ETL pipeline so that each category column is made binary there instead of in the machine learning pipeline. I also made the check for messages that would cause CountVectorizer to throw an error more efficient. Finally, I decided to include 'child_alone' in the classification tasks even though the data includes no positive instances of it.